Next iteration of specifier priority system
This needs to be cleaned up, but now applyTo() is passed a dict of
property being specified keys mapped to boolean values that, if True,
means that that property is being modified by the specifier. Next
part will be adding a helper function to DelayedArguments of modifying
specifiers that will use this dict to decide whether they need to read
an existing value.

All other specifiers just ignore this as they never modify anything.